requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
PyJWT
pydantic
pydantic-settings